补充 11前文件内容

+++ 10-file-system/03-my-mv.js
@@ -3,6 +3,15 @@
 const fs = require('fs'),
       src = process.argv[2],
       dst = process.argv[3];
-fs.renameSync(src,dst);
 
-
+try{
+  if(fs.existsSync(src)){
+    fs.renameSync(src, dst);
+  }else{
+    console.error(err.message);
+    process.exit(1);
+  }
+} catch(err) {
+  console.error(err.message);
+  process.exit(1);
+}
