+++ 10-file-system/04-my-rm.js
@@ -2,7 +2,14 @@
 
 const fs = require('fs'),
       file = process.argv[2];
-fs.unlinkSync(file);
+if(fs.existsSync(file)){
+  if(fs.statSync(file).isFile()){
+    fs.unlinkSync(file);
+  }
+}else{
+  console.error('%s not exist!', file);
+  process.exit(1);
+}
